Separated Directive code from Apache2.pm, these will have to be handled later, after we have most of the rest of the code checked and at least syntax free, and installable by "make install".
Actually testing the directives will be a hurdle left until we are actually
ready to begin the Apache phase of the port.

As far as Apache1 support, It will probably be best to leave it as a separate
part of the build tree. That is another hurdle that will be left to LAST. We aren't
concerned with the Apache1 specific part of the code in any active sense.